Refactored Texture design again, now understanding is in place

+++ src/graphics/PixelData.cpp
@@ -177,7 +177,7 @@ Color PixelData::GetPixelAt(GuiUnit_t aX, GuiUnit_t aY, Color aColor) const
 
         case ColorDepth::RGBA:
             offset = ((aY * GetWidth()) + aX) * 4;
-            result = *reinterpret_cast<const std::uint32_t*>(mpData + offset);
+            result = *reinterpret_cast<const std::uint32_t*>(uintptr_t(mpData + offset));
 //            result.SetRed(mpData[offset + 0]);
 //            result.SetGreen(mpData[offset + 1]);
 //            result.SetBlue(mpData[offset + 2]);
@@ -238,10 +238,10 @@ PixelData& PixelData::SetPixelAt(GuiUnit_t aX, GuiUnit_t aY, Color aColor)
         case ColorDepth::RGBA:
             offset = ((aY * GetWidth()) + aX) * 4;
             if (!mBlend || aColor.GetAlpha() == 255) {
-                *reinterpret_cast<std::uint32_t*>(pdata + offset) = aColor.AsUint();
+                *reinterpret_cast<std::uint32_t*>(uintptr_t(pdata + offset)) = aColor.AsUint();
             }
             else {
-                *reinterpret_cast<std::uint32_t*>(pdata + offset) = Color::Blend(*reinterpret_cast<std::uint32_t*>(pdata + offset), aColor);
+                *reinterpret_cast<std::uint32_t*>(uintptr_t(pdata + offset)) = Color::Blend(*reinterpret_cast<std::uint32_t*>(uintptr_t(pdata + offset)), aColor);
             }
 //            pdata[offset + 0] = aColor.GetRed();
 //            pdata[offset + 1] = aColor.GetGreen();
